// frontend/certifications-ui/src/app/core/api/generated/certificationsApiV1.schemas.ts
export interface CertificationDto {
  certificationId: number;
  contractId: number;
  assessor: string;
  certificationDate: string;
  /** @nullable */
  protocolDate?: string | null;
  /** @nullable */
  prolongationSend?: string | null;
  /** @nullable */
  prolongationReturned?: string | null;
  isCompleted: boolean;
}
// ...
export interface ContractDto {
  contractId: number;
  employeeId: string;
  position: string;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  division?: string | null;
  contractDate: string;
  /** @nullable */
  validTo?: string | null;
  effectiveValidTo: string;
  active: boolean;
  prolongationWarningMonths: number;
  prolongationAlertMonths: number;
  prolongationForYears: number;
  rowVersion: number;
  status: ContractDtoStatus;
}

export interface ContractDetailsDto {
  contract: ContractDto;
  certifications: CertificationDto[];
}
// ...
export type EmployeeDetailsDtoPreferredAdminMode =
  | (typeof EmployeeDetailsDtoPreferredAdminMode)[keyof typeof EmployeeDetailsDtoPreferredAdminMode]
  | null;

export const EmployeeDetailsDtoPreferredAdminMode = {
  MyPage: 'MyPage',
  Administration: 'Administration',
} as const;

export interface EmployeeDetailsDto {
  employeeId: string;
  personalId: string;
  firstName: string;
  /** @nullable */
  middleName?: string | null;
  lastName: string;
  isAdmin: boolean;
  /** @nullable */
  preferredAdminMode?: EmployeeDetailsDtoPreferredAdminMode;
  currentContract?: ContractDetailsDto | null;
}
// ...
export interface UpdateEmployeeRequest {
  personalId: string;
  firstName: string;
  /** @nullable */
  middleName?: string | null;
  lastName: string;
  isAdmin: boolean;
}